client/python/requirements.txt
python-etcd>=0.4.5
urllib3>=1.7.1